requirements.txt
mysql-connector-python==8.0.33
bcrypt==4.0.1
pycryptodome==3.18.0
